feat: add puzzle slice to pages
