+++ packages/edition-slices/src/edition-slices.js
@@ -9,7 +9,8 @@ import {
   SecondaryOneAndFourSlice,
   SecondaryOneSlice,
   SecondaryTwoAndTwoSlice,
-  SecondaryTwoNoPicAndTwoSlice
+  SecondaryTwoNoPicAndTwoSlice,
+  PuzzleSlice
 } from "./slices";
 
 const sliceMap = {
@@ -18,6 +19,7 @@ const sliceMap = {
   LeadOneAndOneSlice,
   LeadOneFullWidthSlice,
   LeadTwoNoPicAndTwoSlice,
+  Puzzle: PuzzleSlice,
   SecondaryFourSlice,
   SecondaryOneAndFourSlice,
   SecondaryOneSlice,
