+++ packages/edition-slices/src/slices/puzzle/proptypes.js
@@ -3,7 +3,11 @@ import PropTypes from "prop-types";
 const propTypes = {
   onPress: PropTypes.func.isRequired,
   slice: PropTypes.shape({
-    image: PropTypes.string,
+    image: PropTypes.shape({
+      crop32: PropTypes.shape({
+        url: PropTypes.string
+      })
+    }),
     title: PropTypes.string,
     url: PropTypes.string
   }).isRequired
