+++ packages/fixture-generator/src/times-image.ts
@@ -1,0 +1,10 @@
+import { Crop, Image } from "./types";
+
+interface TimesImage extends Image {
+  crop11: Crop | null;
+  crop169: Crop | null;
+  crop32: Crop | null;
+  crop45: Crop | null;
+}
+
+export default TimesImage;
